+++ src/Sync/Processor/ObjectSyncProcessor.php
@@ -68,6 +68,7 @@ abstract class ObjectSyncProcessor implements IObjectSyncProcessor {
 	 */
 	final public function process(IObject $object, IDataTransferObject $dto, bool $force = false) {
 		$hook = new HookObject($object);
+
 		// We keep the old data if the object is getting deleted, as there is no "real" DTO available, because
 		// the data has not been delivered...
 		if ($object->getStatus() != IObject::STATUS_TO_DELETE) {
@@ -116,12 +117,14 @@ abstract class ObjectSyncProcessor implements IObjectSyncProcessor {
 				}
 				break;
 			case IObject::STATUS_TO_DELETE:
-				$this->implementation->beforeDeleteILIASObject($hook);
-				$ilias_object = $this->handleDelete($object->getILIASId());
-				if ($ilias_object === null) {
-					throw new ILIASObjectNotFoundException($object);
+				if(!$this->implementation->ignoreDelete($hook)){
+					$this->implementation->beforeDeleteILIASObject($hook);
+					$ilias_object = $this->handleDelete($object->getILIASId());
+					if ($ilias_object === null) {
+						throw new ILIASObjectNotFoundException($object);
+					}
+					$this->implementation->afterDeleteILIASObject($hook->withILIASObject($ilias_object));
 				}
-				$this->implementation->afterDeleteILIASObject($hook->withILIASObject($ilias_object));
 				break;
 			case IObject::STATUS_IGNORED:
 				// Nothing to do here, object is ignored
